[Silver V] Title: 너의 평점은, Time: 104 ms, Memory: 14360 KB -BaekjoonHub

+++ 백준/Silver/25206. 너의 평점은/너의 평점은.java
@@ -11,7 +11,7 @@ public class Main {
         for(int i = 0; i < 20; i++) {
             st = new StringTokenizer(br.readLine(), " ");
             st.nextToken();
-            float score = Float.parseFloat(st.nextToken());
+            double score = Double.parseDouble(st.nextToken());
             String grade = st.nextToken();
             if(!grade.equals("P")) {
                 sum += score;
